+++ projects/summer-fi/helpers/get-automation-cdp-ids.js
